#17 - Implement Register getSignal and setSignal methods

+++ src/cpu.hpp
@@ -14,7 +14,7 @@ public:
 
     ~Register();
 
-    void updateValue(float);
+    void updateValue(int);
 
     float getIntValue();
